+++ .env.sample
@@ -1,2 +1,5 @@
 HOST=localhost
 PORT=3000
+MONGO_USER=usernamefrommongoatlas
+MONGO_PWD=passwordfrommongoatlas
+#mongodb+srv://<username>:<password>@robotdreams.nnmkd7g.mongodb.net/?retryWrites=true&w=majority
